+++ app/src/main/java/com/flytekart/models/request/CreateVariantRequest.java
@@ -26,15 +26,15 @@ public class CreateVariantRequest {
 
     @SerializedName("price")
     @Expose
-    private float price;
+    private double price;
 
     @SerializedName("tax")
     @Expose
-    private float tax;
+    private double tax;
 
     @SerializedName("originalPrice")
     @Expose
-    private float originalPrice;
+    private double originalPrice;
 
     @SerializedName("active")
     @Expose
@@ -72,27 +72,27 @@ public class CreateVariantRequest {
         this.sku = sku;
     }
 
-    public float getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(float price) {
+    public void setPrice(double  price) {
         this.price = price;
     }
 
-    public float getTax() {
+    public double getTax() {
         return tax;
     }
 
-    public void setTax(float tax) {
+    public void setTax(double  tax) {
         this.tax = tax;
     }
 
-    public float getOriginalPrice() {
+    public double getOriginalPrice() {
         return originalPrice;
     }
 
-    public void setOriginalPrice(float originalPrice) {
+    public void setOriginalPrice(double  originalPrice) {
         this.originalPrice = originalPrice;
     }
